Add shows counter and refactor code
