+++ src/index.html
@@ -15,10 +15,15 @@
     <header>
         <nav>
             <h1 class="logo">Movies</h1>
-            <div id="" class="search-bar">
+            <div class="right-nav">
+              <h2 class="shows-counter">Top Movies:
+                <span class="count"></span>
+              </h2>
+              <div id="" class="search-bar">
                 <input id="searchInput" placeholder="Search" type="text" class="" />
                 <span class="" id="searchIcon">🔍</span>
               </div>
+            </div>
         </nav>
     </header>
     <main>
