+++ src/style.css
@@ -61,7 +61,25 @@ nav {
 
 .logo {
     font-size: 2.51rem;
-    color: rgba(214,243,31,.9);;
+    color: rgba(214,243,31,.9);
+}
+
+.right-nav {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.shows-counter {
+    /*border: 1px solid rgba(214,243,31,.9);
+    padding: 5px;
+    border-radius: 1rem;
+    background-color: rgba(215, 243, 31, 0.287);
+    color: rgba(214,243,31,.9);*/
+    padding: 10px 7px 10px 7px;
+    border-radius: 1rem;
+    background-color: rgba(214,243,31,.9);
+    color: black;
 }
 
 .card {
